+++ LANDMarkClassifier/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module imports LANDMark and associated classifiers
+"""
+
+from .LANDMark import LANDMarkClassifier
+
+__all__ = ["LANDMarkClassifier"]
